Create Helping_students
Threads-  A thread is a path of execution within a process. A process can contain multiple threads . A thread of execution is the smallest sequence of programmed instructions that can be managed independently by a scheduler, which is typically a part of the operating system.
Semaphores- A semaphore is a data structure used to help threads work together without interfering with each other. The POSIX standard specifies an interface for semaphores, it is not part of Pthreads, but most UNIXes that implement Pthreads also provide semaphores. POSIX semaphores have type sem_t.(which is used in code also).
Mutex - A mutex is a binary variable whose purpose is to provide locking mechanism. It is used to provide mutual exclusion to a section of code, means only one process can work on a particular code section at a time.